Revert "change the data management to use the react developers convention sta…"
