+++ App.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { createAppContainer } from "react-navigation";
 import { createStackNavigator } from "react-navigation-stack";
 import IndexScreen from "./src/screens/IndexScreen";
+import { Provider } from './src/context/BlogContext';
 import ShowScreen from './src/screens/ShowScreen';
 import CreateScreen from './src/screens/CreateScreen';
 import EditScreen from './src/screens/EditScreen';
-import BlogState from './src/context/blogState';
 
 const navigator = createStackNavigator({
   Index: IndexScreen,
@@ -26,8 +26,8 @@ const App = createAppContainer(navigator)
 
 export default () => {
   return (
-    <BlogState>
+    <Provider>
       <App/>
-    </BlogState>
+    </Provider>
   )
 }
