+++ src/screens/ShowScreen.js
@@ -1,14 +1,12 @@
 import React, {  useContext } from 'react'
 import { StyleSheet, View, Text, TouchableOpacity } from 'react-native';
-//import { Context } from '../context/BlogContext';
+import { Context } from '../context/BlogContext';
 import { Feather } from '@expo/vector-icons'
-import blogContext from '../context/blogContext';
 
 const ShowScreen = ({navigation}) => {
-
-    const { blogPosts } = useContext(blogContext)
+    const { state } = useContext(Context)
     const id = navigation.getParam('id')
-    const blogPost = blogPosts.find((blogPost) => blogPost.id === id)
+    const blogPost = state.find((blogPost) => blogPost.id === id)
 
     return ( 
         <View>
@@ -21,6 +19,7 @@ const ShowScreen = ({navigation}) => {
         </View>
     );
 }
+
 ShowScreen.navigationOptions = ({ navigation }) => {
     return {
         headerRight: () => (
