--- types/index.js
@@ -1,3 +1,0 @@
-export const DELETE_BLOGPOST = 'DELETE_BLOGPOST'
-export const EDIT_BLOGPOST = 'EDIT_BLOGPOST'
-export const GET_BLOGPOST = 'GET_BLOGPOST'
